today some changes
