+++ lib/main.dart
@@ -1,7 +1,7 @@
 
 import 'package:community_ecommerce/themes/light_themes.dart';
-import 'package:community_ecommerce/view/favorite_screen.dart';
-import 'package:community_ecommerce/view/women_top_screen.dart';
+import 'package:community_ecommerce/view/favorite_screen/favorite_screen.dart';
+import 'package:community_ecommerce/view/categories_screen/women_top_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
@@ -27,7 +27,7 @@ class MyApp extends StatelessWidget {
         return GetMaterialApp(
           debugShowCheckedModeBanner: false,
           theme: light,
-          home:  FavoriteScreen(),
+          home:  WomenTopScreen(),
           getPages: AppRoutes.appRoutes(),
 
         );
